requirements.txt
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
opensky-api>=1.3.0
requests>=2.31.0
python-dotenv>=1.0.0
jupyter>=1.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
pydantic>=2.4.0 